skibo tells me that that I didn't apply all of his last patch, and sent me a replacement patch that fixes the problem.  The challenge buffer was not large enough by a factor of 4 (due to my changing the size from 128 to 32, but not u_int8_t to u_int32_t).
MFC after: 1 day
Submitted by: skibo@pacbell.net

+++ sys/dev/wi/wi_hostap.c
@@ -505,7 +505,7 @@ wihap_auth_req(struct wi_softc *sc, struct wi_frame *rxfrm,
 	u_int16_t		seq;
 	u_int16_t		status;
 	int			i, challenge_len;
-	u_int8_t		challenge[32];
+	u_int32_t		challenge[32];
 
 	struct wi_80211_hdr	*resp_hdr;
